requirements.txt
django==5.1.2
djangorestframework==3.15.2
django-filter==24.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pandas==2.2.3
openpyxl==3.1.5
reportlab==4.2.5